refactor(typescript): move applyUpdate next to stripNulls (#3220)
applyUpdate encodes the repository-wide convention that undefined leaves a
column untouched while null clears it. It was living inside the feature voting
module, where the next module needing it would have written its own subtly
different version.

It now sits beside stripNulls, the helper it wraps, with tests of its own.
