+++ apps/backend/src/utils/typescript.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest';
-import { mapWithConcurrency } from './typescript';
+import { applyUpdate, mapWithConcurrency } from './typescript';
 
 describe('mapWithConcurrency', () => {
   it('returns an empty array when given an empty list', async () => {
@@ -66,5 +66,40 @@ describe('mapWithConcurrency', () => {
         return item;
       })
     ).rejects.toThrow('boom');
+  });
+});
+
+describe('applyUpdate', () => {
+  it('should drop a null sent for a field that cannot be cleared', () => {
+    const fields = applyUpdate({ name: null, description: 'kept' }, [
+      'description',
+    ]);
+
+    expect(fields).toEqual({ description: 'kept' });
+  });
+
+  it('should keep a null sent for a clearable field', () => {
+    const fields = applyUpdate({ description: null }, ['description']);
+
+    expect(fields).toEqual({ description: null });
+  });
+
+  // Omitting a field and clearing it are two different intents.
+  it('should leave an omitted clearable field untouched', () => {
+    const fields = applyUpdate(
+      { name: 'kept' } as {
+        name: string;
+        description?: string | null;
+      },
+      ['description']
+    );
+
+    expect('description' in fields).toBe(false);
+  });
+
+  it('should keep the falsy values that are not null', () => {
+    const fields = applyUpdate({ position: 0, active: false, labels: [] }, []);
+
+    expect(fields).toEqual({ position: 0, active: false, labels: [] });
   });
 });
